+++ xmake/modules/target/action/install/unix.lua
@@ -43,7 +43,17 @@ function _install_shared_for_package(target, pkg, outputdir)
             if os.isfile(path.join(outputdir, soname)) then
                 wprint("'%s' already exists in install dir, overwriting it from package(%s).", soname, pkg:name())
             end
-            os.vcp(sopath, outputdir)
+            -- we need reserve symlink
+            -- @see https://github.com/xmake-io/xmake/issues/1582
+            os.vcp(sopath, outputdir, {symlink = true})
+            -- copy real file of symlink
+            if os.islink(sopath) then
+                local realpath = os.readlink(sopath)
+                if not path.is_absolute(realpath) then
+                    realpath = path.absolute(realpath, path.directory(sopath))
+                end
+                os.vcp(realpath, outputdir)
+            end
         end
     end
 end